maintainability: decompose runner/tools.go by responsibility (#792)
tools.go was 778 lines — 22 lines under the 800-line hard cap with an empty
baseline, so the next agent-tool change would have forced a rushed split
mid-feature-PR. Split by the three responsibilities already visible in the
decl layout, moving exact line ranges (no behavior change, no renames, no
public API added for the budget):

- tool_sinks.go: the context-sink/audit plumbing — the dynamicToolContextKey
  keys, ToolMutationAudit / ToolMutationSink, linearGraphQLMutationRejected,
  and the With*/*From sink helpers.
- linear_graphql_proxy.go: the linearGraphQLProxy transport+dispatch surface
  (call/callRaw/dispatch, request building, endpoint/auth helpers, allow-set,
  tool description, response wrapping).
- tools.go keeps the tool-surface construction (DynamicToolSet,
  DynamicToolsForWorkflow, input schemas, current-issue guard wiring) plus the
  HTTP/response helpers shared with the gitea proxy and workpad
  (redactToolSecrets, sanitizeToolErrorBody, readDynamicToolResponseBody,
  dynamicToolFailure/Result, request-timeout).

Sizes: tools.go 778->305, tool_sinks.go 117, linear_graphql_proxy.go 376.
Verified lossless: the set of top-level declarations is byte-identical before
and after the split, the moved bodies diff to zero, and the full -race suite +
file-size-budget test pass. github.go and tracker_client.go follow as separate
PRs per the issue.

Closes #792
